Update script.js to fix a bug when no input is supplied and submit btn is clicked
Update script.js to fix a bug when no input is supplied. When no input is supplied and submit button is clicked, the page doesn't reload and if the user enters data at this point and click submit, nothing happens. 

Now the page reloads if no input is supplied and submit button is clicked.